requirements.txt
streamlit~=1.41.1
PyYAML~=6.0
pandas
streamlit-authenticator
pydeck~=0.8.0
PyYAML~=6.0